refactor: SpringApplication 과 Controller 분리
